Controller Support, Icons, and Sounds
Made it so player 2 can start and choose on their side when the shop opens. Also, added fills for the icons so it's not just that square border behind the icon, so edited the struct and data tables for that. Lastly, added some sounds for confirm, error, and hover. Cool.